Public

Public Interface

Documentation for JuDoc.jl's exported functions

JuDoc.serve — Function.

serve(; clear, verb,  port)

Runs JuDoc in the current directory.

Keyword arguments:

clear=false: whether to remove any existing output directory

verb=false: whether to display messages

port=8000: the port to use for the local server (should pick a number between 8000 and 9000)

single=false: whether to run a single pass or run continuously

source

JuDoc.optimize — Function.

optimize(; prerender, minify)

Does a full pass followed by a pre-rendering and minification step.

prerender=true: whether to pre-render katex and highlight.js (requires node.js)

minify=true: whether to minify output (requires python3 and css_html_js_minify)

source

JuDoc.publish — Function.

publish(; minify=true, prerender=true)

This is a simple wrapper doing a git commit and git push without much fanciness. It assumes the current directory is a git folder. This will work in most simple scenarios (e.g. there's only one person updating the website). In other scenarios you should probably do this manually.

Keyword arguments

prerender=true: prerender javascript before pushing see optimize

minify=true: minify output before pushing see optimize

nopass=false: set this to true if you have already run optimize manually.

source